Update autoloaders for Rails 7
Problem
---------

As detailed here, without this change the vanilla rails app with
`effective_datatables` does not load

https://github.com/Sorcery/sorcery/issues/308

Solution
---------

The changes here include

1. manually download and install initializers for `effective_datatables`
   and `effective_resources` because the `rails
   effective_datatables:install` won't even run.  These can be copied
   directly from the respective github repos.
1. add `require: false` in Gemfile to those to libs
1. wrap setup blocks for initializers in
   `Rails.application.reloader.to_prepare`

Open Questions
---------------

Does this really work?

This gets the console up, but it doesn't do anything to *use* those gems
so it's not clear without a more full implementation if this will
actually fix the problem.

It is a step in the right direction.